requirements.txt
flask
requests
beautifulsoup4
transformers
torch
textblob
python-dotenv
pytesseract
Pillow
wordcloud
youtube_transcript_api
sentencepiece
